app/requirements.txt
Flask==3.1.3
pandas==3.0.1
psycopg2-binary==2.9.11
gunicorn==25.1.0
